Move serial port component to dedicated directory.
This shall simplify isolated testing.

Created a wrapper header to hint the LDF where to find the sources.
