+++ lib/3rd_party_adapters/serial_port/SerialOutputStreamBuffer.cpp
