+++ lib/3rd_party_adapters/serial_port/serial_port.hpp
@@ -1,0 +1,8 @@
+/** \file
+ *
+ * This files serves as hint to the PlatformIO Library Dependency Finder to find the corresponding source file.
+ * The actual interface is defined where it is required/used.
+ */
+
+#pragma once
+#include <serial_port_interface.hpp>
