+++ platformio.ini
@@ -81,4 +81,7 @@ test_filter = test_serial_port
 debug_test = test_serial_port
 lib_deps = 
 	${test_native.lib_deps}
-	3rd_party_adapters
+build_flags = 
+	${test_native.build_flags}
+	-Ilib/board_adapters
+lib_ldf_mode = chain
